Add Filter Todo: Make it working
