+++ src/components/Header.vue
@@ -5,7 +5,12 @@
     <div class="w-1/2 flex items-center flex-shrink-0 text-black">
       <!-- <span class="font-semibold text-xl tracking-tight">Logo</span> -->
       <a href="/">
-        <img class="w-12 ml-2 mt-1" src="media/logo-whats_left.png" alt srcset />
+        <img
+          class="w-12 ml-2 mt-1"
+          src="media/logo-whats_left.png"
+          alt
+          srcset
+        />
       </a>
     </div>
     <div class="w-1/2 flex flex-row-reverse">
@@ -26,8 +31,8 @@
         </svg>
       </a>
       <!-- Filter -->
-      <a
-        href="#"
+      <router-link
+        to="/filter"
         class="inline-block text-sm px-4 py-2 leading-none border rounded text-black border-white hover:border-transparent hover:text-blue-400 hover:bg-white mt-1 lg:mt-0"
       >
         <svg
@@ -40,7 +45,7 @@
             d="M2.3 7.7A1 1 0 012 7V3a1 1 0 011-1h18a1 1 0 011 1v4a1 1 0 01-.3.7L15 14.42V17a1 1 0 01-.3.7l-4 4A1 1 0 019 21v-6.59l-6.7-6.7zM4 4v2.59l6.7 6.7a1 1 0 01.3.71v4.59l2-2V14a1 1 0 01.3-.7L20 6.58V4H4z"
           />
         </svg>
-      </a>
+      </router-link>
       <!-- Settings -->
       <a
         href="#"
